+++ projects/moment/src/adapter/index.ts
@@ -6,7 +6,7 @@ import {
 import {
   DatetimeAdapter,
   MAT_DATETIME_FORMATS
-} from "@mat-datetimepicker/core";
+} from "@nader-eloshaiker/mat-datetimepicker";
 import { MomentDatetimeAdapter } from "./moment-datetime-adapter";
 import { MAT_MOMENT_DATETIME_FORMATS } from "./moment-datetime-formats";